Live visualization (#35)
* Fixed a few bugs in multipeak phasing

* Adaptive alien removal for multipeak

Before phasing on a new peak, it compares the shared-object projected diffraction to the actual measured diffraction (both with a 1-pixel gaussian blur). A difference map is defined as the absolute difference between the two blurred images divided by the blurred projection. A mask is generated from all the voxels where the difference map is greater than twice its own median.
While the peak is being phased, the Fourier constraint is enforced using the measurement outside the mask and the projection inside.

* Moved live viewing to its own CoupledRec method so that iterate() isn't so crowded.

* Added a bunch of comments for multipeak stuff

* Adds feedback from the shared object when applying the Fourier constraint. The amount of feedback depends on that peak's confidence value.

* Updated parameters in config_mp:

Removed:
- adaptive_weights (replaced by adapt_trigger)

Added:
- adapt_trigger: Puts the adaptive weighting in the operation/trigger format
- adapt_power: Basically how harshly bad peaks are punished
- adapt_alien_start: When to start removing aliens
- adapt_alien_threshold: Voxels are masked where the error map exceeds this times its median

Renamed:
- weight_X << mp_weight_X
- adapt_threshold_X << peak_threshold_X

* Updated documentation for multipeak configuration

* Added live display trigger for both single and multipeak.

* Fixed colorscales on multipeak live visualization to match single-peak